Add files via upload
upon startup of the project the html currently available is what is the frontend of project, this can later be changed. the index.html in the resources/static/public/ directory is the frontend of the video feature. please ensure you have both gradle and node js in your system path for this to work.

also ensure you have the required library "pillow" for python for the generation of files.

the build file contains all the necessary dependencies needed for this project.

run both servers concurently alongside the python thumbnail generator.
ktor server - localhost:8080/index
node.js server: localhost:3000.

node.js backend is the "app.js" file